chore: Refactored evaluation script for object detection
